Get-ParameterValues
I ran into a problem where default values were not included in "$PSBoundParameters" (Hence the "Bound" part of it). This util fixes it by combining "$MyInvocation.MyCommands.Parameters" and "$PSBoundParameters". Modified it some more because it skipped over booleans when they were set to false. ($value -ne ($null -as $parameter.Value.ParameterType) is always false since $null=$false for a bool)

This also cleaned up the ugliness of adding items to $PSBoundParameters after checking if they exists in New-DSEntry.