Added yet more support
Added support for Error and AlreadyExists SaveResult's under "PUT" method. I prefer to go case-by-case with New-ServerResponse, this way I can evaluate every response and make sure to forge an appropriate ServerResponse, if need be (As in "backend didn't send a descriptive error message/help to give to the end user").