Cs 1011 users refactoring (#17)
* New-DSUser

Remade DSUsers CRUD to match how I did the entries. New-DSUser is the only CMDlet used to create any type of users. Kept New-DSCustomUser for legacy reasons. Also, now supporting Domain users.

* Get-ParameterValues

Removed null check to be sure I'm keeping every parameters. That's how I plan to use it, so removing nulls didn't make sense. It used to look for null by type, so an empty string "" would match $null for string type and be ignored.

* UserSegments

Rewrote user segments (Account, Profile and Security segments) to be more clean and make use of Get-ParameterValues. Also tried to include all/more fields relative to UserEntity model so if they're needed in the future, we won't need to dig deeper.

* Added yet more support

Added support for Error and AlreadyExists SaveResult's under "PUT" method. I prefer to go case-by-case with New-ServerResponse, this way I can evaluate every response and make sure to forge an appropriate ServerResponse, if need be (As in "backend didn't send a descriptive error message/help to give to the end user").

* [IN PROGRESS]Update-DSUser

CustomSecurities don't show properly in angular, most likely due to missing fields that makes it impossible for the backend to convert CustomSecurityEntity to XML

* Update-DSUser - Privileges tab

Can now update fields in privilege tab. Had to force User.userSecurity.repositoryEntities to an array containing only an empty GUID, not sure that it is correct.

* Update-DSUser

Removed $PSBoundParameters at the start of PROCESS block

* Update-DS User - User Type

Now supports user type Admin, User and ReadOnly. Will do Restricted on request.

* Password

Added password field to New/Update-DSUser.
Added a check to make sure that a password is provided if ServerUserType is Builtin.

* UserTypeString

Sets itself automatically in the backend, no need to manage it myself.